feat: use float64 numbers as args to calc ops

+++ adder/src/lib.rs
@@ -5,7 +5,7 @@ use crate::bindings::exports::calc::basic::add::Guest;
 struct Component;
 
 impl Guest for Component {
-    fn add(a: i32, b: i32) -> i32 {
+    fn add(a: f64, b: f64) -> f64 {
         a + b
     }
 }
