+++ calculator/src/lib.rs
@@ -7,7 +7,7 @@ use crate::bindings::exports::calc::basic::calculate::{Guest, Op};
 struct Component;
 
 impl Guest for Component {
-    fn eval_expression(op: Op, x: i32, y: i32) -> i32 {
+    fn eval_expression(op: Op, x: f64, y: f64) -> f64 {
         match op {
             Op::Add => add(x, y),
             Op::Subtract => subtract(x, y),
